user reg

+++ base/views.py
@@ -3,10 +3,11 @@ from django.db.models import Q
 from django.contrib.auth.decorators import login_required
 from .models import Room,Topic
 from .forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
-
+from django.http import HttpResponse
 # rooms=[
 #     {'id':1,'name':'Lets learn python','topic':'Python','description':'This is a python study room'},
 #     {'id':2,'name':'Lets learn django','topic':'Django','description':'This is a django study room'},
@@ -25,8 +26,11 @@ from django.contrib.auth import authenticate,login,logout
 # ]
 
 def loginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method=='POST':
-     username=request.POST.get('username')
+     username=request.POST.get('username').lower()
      password=request.POST.get('password')
 
      try:
@@ -43,12 +47,31 @@ def loginPage(request):
      else:
             messages.error(request,'Username or password is incorrect')
 
-    context={}
+    context={'page':page}
     return render(request, 'base/login_register.html',context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form=UserCreationForm()
+
+    if request.method == 'POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        
+        else:
+            messages.error(request,'An error occured during registration')
+
+
+
+    return render(request,'base/login_register.html',{'form':form})
 
 
 def home(request):
@@ -71,6 +94,7 @@ def room(request ,pk):
     
    return render(request,"base/room.html" ,context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form=RoomForm()
     if request.method=='POST':
@@ -83,10 +107,13 @@ def createRoom(request):
  
     return render(request,"base/room_form.html",context)
 
-
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+
+    if request.user!=room.host:
+        return HttpResponse('You are not alowed here !!!')
     if request.method=='POST':
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -94,9 +121,14 @@ def updateRoom(request,pk):
             return redirect('home')
     context={'form':form}
     return render(request,"base/room_form.html",context)
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!=room.host:
+        return HttpResponse('You are not alowed here !!!')
+
+
+
     if request.method=='POST':
         room.delete()
         return redirect('home')
